\name{adiv-package}
\alias{adiv-package}
\alias{adiv}
\docType{package}
\title{
\packageTitle{adiv}
}
\description{
Package adiv focuses on the analysis of biodiversity
}
\details{
Package \code{adiv} is dedicated to biodiversity indices used in ecology and conservation biology (see e.g. functions \code{divparam}, \code{optimEH}, \code{QE}). It also treats the concepts of species' originality (e.g. \code{distinctDis}, \code{distinctTree}, \code{distinctUltra}) and of dissimilarities between species and communities (e.g. \code{dissABC}, \code{dissRicotta}, \code{dsimcom}). Focus is placed on species, functional and phylogenetic diversity and on the apportionment of biodiversity across spatial and temporal scales (see for notably functions \code{EqRao}, \code{EqRS}, \code{EqRSintra}). Ordination methods are also available to depict in detail biodiversity patterns in space and/or time (e.g. \code{crossdpcoa}, \code{dspca}, \code{evoCA}, \code{evoNSCA}, \code{evopca}).
}
\author{
\packageAuthor{adiv}

Maintainer: \packageMaintainer{adiv}
}
